Refactor: standardize all component file paths to lowercase for consistency ♻️
- Renamed component files (e.g., `Form`, `PageCard`, `Modal`) to lowercase.
- Updated all imports, stubs, and utility references to reflect the new paths.
