+++ resources/js/inertia/auth/User.ts
@@ -1,6 +1,5 @@
 export interface User {
-  first_name: string;
-  last_name: string;
+  first_name?: string;
+  last_name?: string;
   email: string;
-  fcm_token?: string;
 }
