+++ resources/js/inertia/components/show/image-preview.tsx
