+++ resources/js/inertia/components/show/small-text-field.tsx
