+++ src/Settings/Attributes/CubeNumeric.php
@@ -173,7 +173,7 @@ class CubeNumeric extends CubeAttribute implements HasFakeMethod,
             $this->labelNaming(),
             "{$modelVariable}{$nullable}.{$this->name}",
             [
-                new TsImportString("SmallTextField", "@/components/Show/SmallTextField")
+                new TsImportString("SmallTextField", "@/components/show/small-text-field")
             ]
         );
     }
