+++ src/Settings/Attributes/CubeStringable.php
@@ -207,7 +207,7 @@ class CubeStringable extends CubeAttribute implements
             $this->labelNaming(),
             "{$modelVariable}{$nullable}.{$this->name}",
             [
-                new TsImportString("SmallTextField", "@/components/Show/SmallTextField")
+                new TsImportString("SmallTextField", "@/components/show/small-text-field")
             ]
         );
     }
