+++ src/Stub/stubs/Web/InertiaReact/Pages/Form.stub
@@ -1,7 +1,7 @@
 import { useForm } from "@inertiajs/react";
 import { FormEvent } from "react";
-import PageCard from "@/components/ui/PageCard";
-import Form from "@/components/form/Form";
+import PageCard from "@/components/ui/page-card";
+import Form from "@/components/form/form";
 {{imports}}
 
 const {{component_name}} = ({{component_props}}) => {
