+++ src/Stub/stubs/Web/InertiaReact/Pages/UserDetails.stub
@@ -1,6 +1,6 @@
-import Form from "@/components/form/Form";
+import Form from "@/components/form/form";
 import Input from "@/components/form/fields/input";
-import PageCard from "@/components/ui/PageCard";
+import PageCard from "@/components/ui/page-card";
 import User from "@/Models/User";
 import { useForm } from "@inertiajs/react";
 import { FormEvent, useState } from "react";
